requirements.txt
#pip install -r requirements.txt
holdem-calc
python-dotenv